+++ src/components/manage/Blocks/Block/Edit.jsx
@@ -1,6 +1,6 @@
 /**
- * Edit tile.
- * @module components/manage/Tiles/Tile/Edit
+ * Edit block.
+ * @module components/manage/Blocks/Block/Edit
  */
 
 import React, { Component } from 'react';
@@ -9,7 +9,7 @@ import { compose } from 'redux';
 import { DragSource, DropTarget } from 'react-dnd';
 import { findDOMNode } from 'react-dom';
 import { defineMessages, injectIntl } from 'react-intl';
-import { tiles } from '~/config';
+import { blocks } from '~/config';
 import { Button } from 'semantic-ui-react';
 import includes from 'lodash/includes';
 import cx from 'classnames';
@@ -86,7 +86,7 @@ const itemTarget = {
 };
 
 /**
- * Edit tile class.
+ * Edit block class.
  * @class Edit
  * @extends Component
  */
@@ -111,31 +111,31 @@ class Edit extends Component {
 
   componentDidMount() {
     const { type } = this.props;
-    const tileHasOwnFocusManagement =
-      tiles.tilesConfig?.[type]?.['tileHasOwnFocusManagement'] || null;
+    const blockHasOwnFocusManagement =
+      blocks.blocksConfig?.[type]?.['blockHasOwnFocusManagement'] || null;
     if (
-      !tileHasOwnFocusManagement &&
+      !blockHasOwnFocusManagement &&
       this.props.selected &&
-      this.tileNode.current
+      this.blockNode.current
     ) {
-      this.tileNode.current.focus();
+      this.blockNode.current.focus();
     }
   }
 
   componentWillReceiveProps(nextProps) {
     const { type } = this.props;
-    const tileHasOwnFocusManagement =
-      tiles.tilesConfig?.[type]?.['tileHasOwnFocusManagement'] || null;
+    const blockHasOwnFocusManagement =
+      blocks.blocksConfig?.[type]?.['blockHasOwnFocusManagement'] || null;
     if (
-      !tileHasOwnFocusManagement &&
+      !blockHasOwnFocusManagement &&
       nextProps.selected &&
-      this.tileNode.current
+      this.blockNode.current
     ) {
-      this.tileNode.current.focus();
+      this.blockNode.current.focus();
     }
   }
 
-  tileNode = React.createRef();
+  blockNode = React.createRef();
 
   /**
    * Render method.
@@ -152,9 +152,9 @@ class Edit extends Component {
       connectDragPreview,
     } = this.props;
 
-    const Tile = tiles.tilesConfig?.[type]?.['edit'] || null;
-    const tileHasOwnFocusManagement =
-      tiles.tilesConfig?.[type]?.['tileHasOwnFocusManagement'] || null;
+    const Block = blocks.blocksConfig?.[type]?.['edit'] || null;
+    const blockHasOwnFocusManagement =
+      blocks.blocksConfig?.[type]?.['blockHasOwnFocusManagement'] || null;
 
     const hideHandler =
       this.props.data['@type'] === 'text' &&
@@ -179,29 +179,29 @@ class Edit extends Component {
                 <Icon className="drag handle" name={dragSVG} size="18px" />
               </div>,
             )}
-          {Tile !== null ? (
+          {Block !== null ? (
             <div
               role="presentation"
               onClick={() => this.props.onSelectTile(this.props.tile)}
               onKeyDown={
-                !tileHasOwnFocusManagement
+                !blockHasOwnFocusManagement
                   ? e =>
                       this.props.handleKeyDown(
                         e,
                         this.props.index,
                         this.props.tile,
-                        this.tileNode.current,
+                        this.blockNode.current,
                       )
                   : null
               }
               className={cx(`tile ${type}`, { selected: this.props.selected })}
               style={{ outline: 'none' }}
-              ref={this.tileNode}
+              ref={this.blockNode}
               // The tabIndex is required for the keyboard navigation
               /* eslint-disable jsx-a11y/no-noninteractive-tabindex */
-              tabIndex={!tileHasOwnFocusManagement ? -1 : null}
+              tabIndex={!blockHasOwnFocusManagement ? -1 : null}
             >
-              <Tile {...this.props} tileNode={this.tileNode} />
+              <Block {...this.props} blockNode={this.blockNode} />
             </div>
           ) : (
             <div
@@ -211,12 +211,12 @@ class Edit extends Component {
                   e,
                   this.props.index,
                   this.props.tile,
-                  this.tileNode.current,
+                  this.blockNode.current,
                 )
               }
               className={cx(`tile ${type}`, { selected: this.props.selected })}
               style={{ outline: 'none' }}
-              ref={this.tileNode}
+              ref={this.blockNode}
               // The tabIndex is required for the keyboard navigation
               tabIndex={-1}
             >
@@ -225,7 +225,7 @@ class Edit extends Component {
               })}
             </div>
           )}
-          {selected && !includes(tiles.requiredTiles, type) && (
+          {selected && !includes(blocks.requiredBlocks, type) && (
             <Button
               icon
               basic