+++ src/components/manage/Blocks/HeroImageLeft/Edit.jsx
@@ -335,7 +335,7 @@ class Edit extends Component {
                 ) {
                   this.props.onFocusPreviousTile(
                     this.props.tile,
-                    this.props.tileNode.current,
+                    this.props.blockNode.current,
                   );
                 }
               }}
@@ -385,7 +385,7 @@ class Edit extends Component {
                 if (currentCursorPosition === blockLength) {
                   this.props.onFocusNextTile(
                     this.props.tile,
-                    this.props.tileNode.current,
+                    this.props.blockNode.current,
                   );
                 }
               }}