+++ src/helpers/index.js
@@ -25,10 +25,10 @@ export {
   getLayoutFieldname,
 } from '@plone/volto/helpers/Content/Content';
 export {
-  getTilesFieldname,
-  getTilesLayoutFieldname,
-  hasTilesData,
-} from '@plone/volto/helpers/Tiles/Tiles';
+  getBlocksFieldname,
+  getBlocksLayoutFieldname,
+  hasBlocksData,
+} from '@plone/volto/helpers/Blocks/Blocks';
 export BodyClass from '@plone/volto/helpers/BodyClass/BodyClass';
 export ScrollToTop from '@plone/volto/helpers/ScrollToTop/ScrollToTop';
 export {